Extract authentication related code to module
