+++ src/ringgo/mod.rs
@@ -1,2 +1,3 @@
+pub mod authentication;
 pub mod errors;
 pub mod url_helpers;
